Move the docket hearing row into its own component
