+++ client/app/hearings/TextareaContainer.jsx
@@ -1,7 +1,6 @@
 import React from 'react';
 import PropTypes from 'prop-types';
 import { connect } from 'react-redux';
-import { getStateProperty } from './utils';
 import * as Actions from './actions/Dockets';
 
 export class TextareaContainer extends React.Component {
@@ -21,22 +20,14 @@ export class TextareaContainer extends React.Component {
   }
 }
 
-const mapStateToProps = (state, ownProps) => {
-  return {
-    [ownProps.id]: getStateProperty(state, 'docket_note', ownProps.id)
-  };
-};
-
 const mapDispatchToProps = (dispatch) => ({
   updateTextarea: (actionName, prop, value) => {
-    const action = Actions[actionName];
-
-    dispatch(action(prop, value));
+    dispatch(Actions[actionName](prop, value));
   }
 });
 
 export default connect(
-  mapStateToProps,
+  null,
   mapDispatchToProps
 )(TextareaContainer);
